migrated the unit test to spock, fixed the name
